✨ feat: event.comment 관련 파일 이동 및 import 수정
