+++ src/main/java/com/mulmeong/comment/read/application/FeedCommentService.java
@@ -2,9 +2,9 @@ package com.mulmeong.comment.read.application;
 
 import com.mulmeong.comment.read.common.utils.CursorPage;
 import com.mulmeong.comment.read.dto.out.FeedCommentResponseDto;
-import com.mulmeong.event.FeedCommentCreateEvent;
-import com.mulmeong.event.FeedCommentDeleteEvent;
-import com.mulmeong.event.FeedCommentUpdateEvent;
+import com.mulmeong.event.comment.FeedCommentCreateEvent;
+import com.mulmeong.event.comment.FeedCommentDeleteEvent;
+import com.mulmeong.event.comment.FeedCommentUpdateEvent;
 
 public interface FeedCommentService {
     void createFeedComment(FeedCommentCreateEvent message);
