+++ src/main/java/com/mulmeong/comment/read/application/FeedCommentServiceImpl.java
@@ -7,13 +7,11 @@ import com.mulmeong.comment.read.dto.out.FeedCommentResponseDto;
 import com.mulmeong.comment.read.entity.FeedComment;
 import com.mulmeong.comment.read.infrsatructure.FeedCommentRepository;
 import com.mulmeong.comment.read.infrsatructure.FeedCommentRepositoryCustom;
-import com.mulmeong.event.FeedCommentCreateEvent;
-import com.mulmeong.event.FeedCommentDeleteEvent;
-import com.mulmeong.event.FeedCommentUpdateEvent;
+import com.mulmeong.event.comment.FeedCommentCreateEvent;
+import com.mulmeong.event.comment.FeedCommentDeleteEvent;
+import com.mulmeong.event.comment.FeedCommentUpdateEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
