+++ src/main/java/com/mulmeong/comment/read/application/FeedRecommentServiceImpl.java
@@ -8,9 +8,9 @@ import com.mulmeong.comment.read.entity.FeedComment;
 import com.mulmeong.comment.read.entity.FeedRecomment;
 import com.mulmeong.comment.read.infrsatructure.FeedRecommentRepository;
 import com.mulmeong.comment.read.infrsatructure.FeedRecommentRepositoryCustom;
-import com.mulmeong.event.FeedRecommentCreateEvent;
-import com.mulmeong.event.FeedRecommentDeleteEvent;
-import com.mulmeong.event.FeedRecommentUpdateEvent;
+import com.mulmeong.event.comment.FeedRecommentCreateEvent;
+import com.mulmeong.event.comment.FeedRecommentDeleteEvent;
+import com.mulmeong.event.comment.FeedRecommentUpdateEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.mongodb.core.MongoTemplate;
