+++ src/main/java/com/mulmeong/comment/read/application/KafkaConsumer.java
@@ -1,6 +1,6 @@
 package com.mulmeong.comment.read.application;
 
-import com.mulmeong.event.*;
+import com.mulmeong.event.comment.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -17,21 +17,21 @@ public class KafkaConsumer {
     private final ShortsRecommentService shortsRecommentService;
 
     //feed comment
-    @KafkaListener(topics = "feed-comment-create",
+    @KafkaListener(topics = "feed-comment-created",
                    groupId = "comment-read",
                    containerFactory = "feedCommentCreateListener")
     public void createFeedComment(FeedCommentCreateEvent message) {
         feedCommentService.createFeedComment(message);
     }
 
-    @KafkaListener(topics = "feed-comment-update",
+    @KafkaListener(topics = "feed-comment-updated",
                    groupId = "comment-read",
                    containerFactory = "feedCommentUpdateListener")
     public void updateFeedComment(FeedCommentUpdateEvent message) {
         feedCommentService.updateFeedComment(message);
     }
 
-    @KafkaListener(topics = "feed-comment-delete",
+    @KafkaListener(topics = "feed-comment-deleted",
                    groupId = "comment-read",
                    containerFactory = "feedCommentDeleteListener")
     public void deleteFeedComment(FeedCommentDeleteEvent message) {
@@ -39,21 +39,21 @@ public class KafkaConsumer {
     }
 
     //feed recomment
-    @KafkaListener(topics = "feed-recomment-create",
+    @KafkaListener(topics = "feed-recomment-created",
                    groupId = "comment-read",
                    containerFactory = "feedRecommentCreateListener")
     public void createFeedRecomment(FeedRecommentCreateEvent message) {
         feedRecommentService.createFeedRecomment(message);
     }
 
-    @KafkaListener(topics = "feed-recomment-update",
+    @KafkaListener(topics = "feed-recomment-updated",
                    groupId = "comment-read",
                    containerFactory = "feedRecommentUpdateListener")
     public void updateFeedRecomment(FeedRecommentUpdateEvent message) {
         feedRecommentService.updateFeedRecomment(message);
     }
 
-    @KafkaListener(topics = "feed-recomment-delete",
+    @KafkaListener(topics = "feed-recomment-deleted",
                    groupId = "comment-read",
                    containerFactory = "feedRecommentDeleteListener")
     public void deleteFeedRecomment(FeedRecommentDeleteEvent message) {
@@ -61,21 +61,21 @@ public class KafkaConsumer {
     }
 
     //shorts comment
-    @KafkaListener(topics = "shorts-comment-create",
+    @KafkaListener(topics = "shorts-comment-created",
                    groupId = "comment-read",
                    containerFactory = "shortsCommentCreateListener")
     public void createShortsComment(ShortsCommentCreateEvent message) {
         shortsCommentService.createShortsComment(message);
     }
 
-    @KafkaListener(topics = "shorts-comment-update",
+    @KafkaListener(topics = "shorts-comment-updated",
                    groupId = "comment-read",
                    containerFactory = "shortsCommentUpdateListener")
     public void updateShortsComment(ShortsCommentUpdateEvent message) {
         shortsCommentService.updateShortsComment(message);
     }
 
-    @KafkaListener(topics = "shorts-comment-delete",
+    @KafkaListener(topics = "shorts-comment-deleted",
                    groupId = "comment-read",
                    containerFactory = "shortsCommentDeleteListener")
     public void deleteShortsComment(ShortsCommentDeleteEvent message) {
@@ -83,21 +83,21 @@ public class KafkaConsumer {
     }
 
     //shorts recomment
-    @KafkaListener(topics = "shorts-recomment-create",
+    @KafkaListener(topics = "shorts-recomment-created",
                    groupId = "comment-read",
                    containerFactory = "shortsRecommentCreateListener")
     public void createShortsRecomment(ShortsRecommentCreateEvent message) {
         shortsRecommentService.createShortsRecomment(message);
     }
 
-    @KafkaListener(topics = "shorts-recomment-update",
+    @KafkaListener(topics = "shorts-recomment-updated",
                    groupId = "comment-read",
                    containerFactory = "shortsRecommentUpdateListener")
     public void updateShortsRecomment(ShortsRecommentUpdateEvent message) {
         shortsRecommentService.updateShortsRecomment(message);
     }
 
-    @KafkaListener(topics = "shorts-recomment-delete",
+    @KafkaListener(topics = "shorts-recomment-deleted",
                    groupId = "comment-read",
                    containerFactory = "feedRecommentDeleteListener")
     public void deleteShortsRecomment(ShortsRecommentDeleteEvent message) {
