+++ src/main/java/com/mulmeong/comment/read/application/ShortsCommentServiceImpl.java
@@ -3,15 +3,13 @@ package com.mulmeong.comment.read.application;
 import com.mulmeong.comment.read.common.exception.BaseException;
 import com.mulmeong.comment.read.common.response.BaseResponseStatus;
 import com.mulmeong.comment.read.common.utils.CursorPage;
-import com.mulmeong.comment.read.dto.out.FeedCommentResponseDto;
 import com.mulmeong.comment.read.dto.out.ShortsCommentResponseDto;
-import com.mulmeong.comment.read.entity.FeedComment;
 import com.mulmeong.comment.read.entity.ShortsComment;
 import com.mulmeong.comment.read.infrsatructure.ShortsCommentRepository;
 import com.mulmeong.comment.read.infrsatructure.ShortsCommentRepositoryCustom;
-import com.mulmeong.event.ShortsCommentCreateEvent;
-import com.mulmeong.event.ShortsCommentDeleteEvent;
-import com.mulmeong.event.ShortsCommentUpdateEvent;
+import com.mulmeong.event.comment.ShortsCommentCreateEvent;
+import com.mulmeong.event.comment.ShortsCommentDeleteEvent;
+import com.mulmeong.event.comment.ShortsCommentUpdateEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.mongodb.core.MongoTemplate;
