+++ src/main/java/com/mulmeong/comment/read/application/ShortsRecommentServiceImpl.java
@@ -3,16 +3,14 @@ package com.mulmeong.comment.read.application;
 import com.mulmeong.comment.read.common.exception.BaseException;
 import com.mulmeong.comment.read.common.response.BaseResponseStatus;
 import com.mulmeong.comment.read.common.utils.CursorPage;
-import com.mulmeong.comment.read.dto.out.FeedRecommentResponseDto;
 import com.mulmeong.comment.read.dto.out.ShortsRecommentResponseDto;
-import com.mulmeong.comment.read.entity.FeedComment;
-import com.mulmeong.comment.read.entity.FeedRecomment;
 import com.mulmeong.comment.read.entity.ShortsComment;
 import com.mulmeong.comment.read.entity.ShortsRecomment;
-import com.mulmeong.comment.read.infrsatructure.ShortsCommentRepositoryCustom;
 import com.mulmeong.comment.read.infrsatructure.ShortsRecommentRepository;
 import com.mulmeong.comment.read.infrsatructure.ShortsRecommentRepositoryCustom;
-import com.mulmeong.event.*;
+import com.mulmeong.event.comment.ShortsRecommentCreateEvent;
+import com.mulmeong.event.comment.ShortsRecommentDeleteEvent;
+import com.mulmeong.event.comment.ShortsRecommentUpdateEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.mongodb.core.MongoTemplate;
