--- src/main/java/com/mulmeong/event/FeedCommentCreateEvent.java
@@ -1,40 +1,0 @@
-package com.mulmeong.event;
-
-import com.mulmeong.comment.read.entity.FeedComment;
-import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Data
-@Getter
-@NoArgsConstructor
-public class FeedCommentCreateEvent {
-
-    private String feedUuid;
-    private String memberUuid;
-    private String commentUuid;
-    private boolean status;
-    private String content;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-    private Integer likeCount;
-    private Integer dislikeCount;
-
-    public FeedComment toEntity() {
-        return FeedComment.builder()
-                .feedUuid(feedUuid)
-                .memberUuid(memberUuid)
-                .commentUuid(commentUuid)
-                .content(content)
-                .status(status)
-                .createdAt(createdAt)
-                .updatedAt(updatedAt)
-                .likeCount(likeCount)
-                .dislikeCount(dislikeCount)
-                .recommentCount(0)
-                .build();
-    }
-
-}
