+++ src/main/java/com/mulmeong/event/comment/FeedCommentCreateEvent.java
@@ -1,7 +1,6 @@
-package com.mulmeong.event;
+package com.mulmeong.event.comment;
 
-
-import com.mulmeong.comment.read.entity.FeedRecomment;
+import com.mulmeong.comment.read.entity.FeedComment;
 import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,27 +10,30 @@ import java.time.LocalDateTime;
 @Data
 @Getter
 @NoArgsConstructor
-public class FeedRecommentCreateEvent {
+public class FeedCommentCreateEvent {
 
+    private String feedUuid;
     private String memberUuid;
     private String commentUuid;
-    private String recommentUuid;
+    private boolean status;
     private String content;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
     private Integer likeCount;
     private Integer dislikeCount;
 
-    public FeedRecomment toEntity() {
-        return FeedRecomment.builder()
+    public FeedComment toEntity() {
+        return FeedComment.builder()
+                .feedUuid(feedUuid)
                 .memberUuid(memberUuid)
                 .commentUuid(commentUuid)
-                .recommentUuid(recommentUuid)
                 .content(content)
+                .status(status)
                 .createdAt(createdAt)
                 .updatedAt(updatedAt)
                 .likeCount(likeCount)
                 .dislikeCount(dislikeCount)
+                .recommentCount(0)
                 .build();
     }
