+++ src/main/java/com/mulmeong/event/comment/ShortsRecommentCreateEvent.java
@@ -1,4 +1,5 @@
-package com.mulmeong.event;
+package com.mulmeong.event.comment;
+
 
 import com.mulmeong.comment.read.entity.ShortsRecomment;
 import lombok.Data;
@@ -10,23 +11,28 @@ import java.time.LocalDateTime;
 @Data
 @Getter
 @NoArgsConstructor
-public class ShortsRecommentUpdateEvent {
+public class ShortsRecommentCreateEvent {
 
+    private String memberUuid;
+    private String commentUuid;
     private String recommentUuid;
     private String content;
+    private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
+    private Integer likeCount;
+    private Integer dislikeCount;
 
-    public ShortsRecomment toEntity(ShortsRecomment shortsRecomment) {
+    public ShortsRecomment toEntity() {
         return ShortsRecomment.builder()
-                .id(shortsRecomment.getId())
-                .memberUuid(shortsRecomment.getMemberUuid())
-                .commentUuid(shortsRecomment.getCommentUuid())
+                .memberUuid(memberUuid)
+                .commentUuid(commentUuid)
                 .recommentUuid(recommentUuid)
                 .content(content)
-                .createdAt(shortsRecomment.getCreatedAt())
+                .createdAt(createdAt)
                 .updatedAt(updatedAt)
-                .likeCount(shortsRecomment.getLikeCount())
-                .dislikeCount(shortsRecomment.getDislikeCount())
+                .likeCount(likeCount)
+                .dislikeCount(dislikeCount)
                 .build();
     }
+
 }
