+++ src/main/java/com/mulmeong/event/comment/ShortsRecommentDeleteEvent.java
@@ -1,4 +1,4 @@
-package com.mulmeong.event;
+package com.mulmeong.event.comment;
 
 import lombok.Data;
 import lombok.Getter;
